fixed linux error due to sizeof(long). compilled linux .so

+++ include/compress.h
@@ -12,33 +12,33 @@ static const char V0[] = "ANALIZER1.0";
 static const char V1[] = "ANALIZER1.1";
 static const char V2[] = "ANALIZER1.2";
 
-int DefineVersion(const char * str);
+int32_t DefineVersion(const char * str);
 
 typedef struct {
-    unsigned short year;
-    unsigned short month;
-    unsigned short dayOfWeek;
-    unsigned short day;
-    unsigned short hour;
-    unsigned short minute;
-    unsigned short second;
-    unsigned short milliseconds;
+    uint16_t year;
+    uint16_t month;
+    uint16_t dayOfWeek;
+    uint16_t day;
+    uint16_t hour;
+    uint16_t minute;
+    uint16_t second;
+    uint16_t milliseconds;
 } Time;
 typedef struct {
-    int type;
+    int32_t type;
     char name[128];
     char comment[128];
     char unit[128];
     Time time;
-    int nPoints;
+    int32_t nPoints;
     double tMin, tMax, yMin, delta; // omitted in output
     unsigned char data[1];
 }CombiscopeHistogram;
 
-static const int SIGNAL_HEADER_SIZE = 408;
+static const int32_t SIGNAL_HEADER_SIZE = 408;
 
 typedef struct {
-    unsigned short Vertex[2];
+    uint16_t Vertex[2];
 }CompressionGraph;
 typedef struct {
     unsigned char Bits[32];
@@ -46,8 +46,8 @@ typedef struct {
 }Code;
 class Knot{
 public:
-    int weight;
-    int parent;
+    int32_t weight;
+    int32_t parent;
 
     Knot() : weight(0), parent(255){};
 };
@@ -65,7 +65,7 @@ public:
     unsigned char* data;
     const size_t size;
 
-    explicit CompressedRLE(const int size) : size(size){
+    explicit CompressedRLE(const int32_t size) : size(size){
         data = new unsigned char [size];
     };
     ~CompressedRLE(){
@@ -74,11 +74,11 @@ public:
 };
 
 union LongFlip{
-    long asLong;
+    int32_t asLong;
     unsigned char asChar[4];
 };
 union LongToDouble {
-    long asLong[2];
+    int32_t asLong[2];
     double asDouble;
 };
 
@@ -86,7 +86,7 @@ static std::vector<std::thread> workers;
 static std::vector<CompressedHoff> tasks;
 static std::mutex lockIn;
 static std::mutex lockOut;
-static unsigned int reqCount;
+static uint32_t reqCount;
 static char* req;
 static Out out = {
         0,
@@ -94,21 +94,21 @@ static Out out = {
 };
 static char* currentOutPos = nullptr;
 
-Out parseSHT(const char* path, unsigned int reqCount, char* requests);
+Out parseSHT(const char* path, uint32_t reqCount, char* requests);
 
 CompressedRLE* DecompressHoffman(const CompressedHoff* compressed);
 CombiscopeHistogram* DecompressRLE(const CompressedRLE* compressed);
 void appendOut(const CombiscopeHistogram*);
 
-Out packSHT(int signalCount, const char* headers, const char* data);
-Out packADC(int signalCount, const char* headers, const char* data);
-CompressedRLE* compressRLE(const CombiscopeHistogram* uncompressed, const int size);
+Out packSHT(int32_t signalCount, const char* headers, const char* data);
+Out packADC(int32_t signalCount, const char* headers, const char* data);
+CompressedRLE* compressRLE(const CombiscopeHistogram* uncompressed, const int32_t size);
 CompressedHoff compressHoffman(const CompressedRLE* uncompressed);
 
 static Out* outs;
 //static size_t curr = 0;
 
-int innerTest(int n);
+int32_t innerTest(int32_t n);
 
 void innerFreeOut();
